server/voting_node/requirements.txt
Flask>=3.0.0
requests>=2.31.0
PyJWT>=2.8.0
bcrypt>=4.1.0
cryptography>=41.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
python-dotenv>=1.0.0
